added some components
