+++ src/Parts.js
@@ -1,0 +1,40 @@
+import React from 'react';
+import slugify from 'slugify';
+
+export default function Parts (props) {
+  const features = Object.keys(props.features).map((feature, idx) => {
+    const featureHash = feature + '-' + idx;
+    //should probably be another component
+    const options = props.features[feature].map(item => {
+      const itemHash = slugify(JSON.stringify(item));
+      return (
+        <div key={itemHash} className="feature__item">
+          <input
+            type="radio"
+            id={itemHash}
+            className="feature__option"
+            name={slugify(feature)}
+            checked={item.name === props.selected[feature].name}
+            onChange={e => props.handleUpdate(feature, item)}
+            />
+            <label htmlFor={itemHash} className="feature__label">
+              {item.name} ({props.currencyFormat.format(item.cost)})
+            </label>
+        </div>
+      );
+    })
+
+    return (
+      <fieldset className="feature" key={featureHash}>
+        <legend className="feature__name">
+          <h3>{feature}</h3>
+        </legend>
+        {options}
+      </fieldset>
+    );
+  });
+  
+  return (
+  <div>{features}</div>
+  );
+}
