+++ src/index.js
@@ -2,8 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
-import features from './FEATURES';
 
 
 
-ReactDOM.render(<App features={features}/>, document.getElementById('root'));
+ReactDOM.render(<App />, document.getElementById('root'));
